Moved the function and the localVariable to the main page
